feat: add authentication

+++ src/graphql/users/resolvers.js
@@ -1,3 +1,5 @@
+import { AuthenticationError } from "apollo-server"
+
 // Query resolvers
 async function user(_, { id }, { dataSources }) {
     return await dataSources.usersApi.getUser(id)
@@ -12,11 +14,18 @@ async function createUser(_, { data }, { dataSources }) {
     return dataSources.usersApi.createUser(data)
 }
 
-async function updateUser(_, { id, data }, { dataSources }) {
+async function updateUser(_, { id, data }, { dataSources, loggedUserId }) {
+    if (!loggedUserId) throw new AuthenticationError("You must be logged!")
+    if (id !== loggedUserId) {
+        throw new AuthenticationError("You cannot update this user!")
+    }
     return dataSources.usersApi.updateUser(id, data)
 }
 
-async function deleteUser(_, { id }, { dataSources }) {
+async function deleteUser(_, { id }, { dataSources, loggedUserId }) {
+    if (id !== loggedUserId) {
+        throw new AuthenticationError("You cannot delete this user!")
+    }
     return dataSources.usersApi.deleteUser(id)
 }
